+++ R/friends.R
@@ -77,7 +77,7 @@ get_friends <- function(users,
                         token = NULL) {
   
 
-  stopifnot(is.vector(users), is_n(n))
+  stopifnot(is_n(n))
 
   results <- lapply(users, get_friends_user, 
     n = n, 
@@ -94,7 +94,7 @@ get_friends <- function(users,
 
 get_friends_user <- function(user, token, n = 5000, parse = TRUE) {
   params <- list(stringify_ids = TRUE)
-  params[[.id_type(user)]] <- user
+  params[[user_type(user)]] <- user
   
   results <- TWIT_paginate_cursor(token, "/1.1/friends/ids", params, 
     page_size = 5000,
@@ -116,8 +116,7 @@ get_friends_user <- function(user, token, n = 5000, parse = TRUE) {
 #' Gets information on friendship between authenticated user and up
 #' to 100 other users.
 #'
-#' @inheritParams lookup_users
-#' @param user Screen name or user id of target user.
+#' @inheritParams get_timeline
 #' @return Data frame converted form returned JSON object. If parse is
 #'   not true, the HTTP response object is returned instead.
 #' @family friends
@@ -127,7 +126,7 @@ my_friendships <- function(user,
                            parse = TRUE,
                            token = NULL) {
   params <- list()
-  params[[.id_type(user)]] <- paste0(user, collapse = ",")
+  params[[user_type(user)]] <- paste0(user, collapse = ",")
   TWIT_get(token, "/1.1/friendships/lookup", params, parse = parse)
 }
 
@@ -166,8 +165,8 @@ lookup_friendships_ <- function(source,
   stopifnot(is.atomic(source), is.atomic(target))
 
   params <- list()
-  params[[paste0("source_", .id_type(source))]] <- source
-  params[[paste0("target_", .id_type(target))]] <- target
+  params[[paste0("source_", user_type(source, "source"))]] <- source
+  params[[paste0("target_", user_type(target, "target"))]] <- target
 
   f <- TWIT_get(token, "/1.1/friendships/show", params, parse = parse)
   if (parse) {