+++ tests/testthat/test-mentions.R
@@ -1,5 +1,5 @@
 test_that("mentions returns tweets data", {
-  x <- get_mentions()
+  suppressMessages(x <- get_mentions())
 
   expect_s3_class(x, "tbl_df")
   expect_gt(nrow(x), 0)