+++ tests/testthat/test-post-user.R
@@ -1,20 +1,20 @@
-test_that("post_follow works", {
+test_that("can follow and unfollow", {
+  skip("requires manual testing")
   pf <- post_follow("BarackObama")
   expect_equal(httr::status_code(pf), 200L)
-})
-
-test_that("post_destroy works", {
+  
   pf <- post_follow("BarackObama", destroy = TRUE)
   expect_equal(httr::status_code(pf), 200L)
 })
 
-test_that("post_follow works with notifications", {
+test_that("can follow and unfollow with notifications", {
+  skip("requires manual testing")
   pf <- post_follow("BarackObama", notify = TRUE)
   expect_equal(httr::status_code(pf), 200L)
+  
   pf <- post_follow("BarackObama", destroy = TRUE)
   expect_equal(httr::status_code(pf), 200L)
 })
-
 
 test_that("Muting  #467", {
   pf <- post_follow("hlynur", destroy = TRUE, mute = TRUE, notify = TRUE)